+++ .gitignore
@@ -10,9 +10,6 @@
 *.sw?
 
 vercel.local.json
-.env
-.env*.local
-
 
 # dependencies
 /node_modules
