+++ messages/en1.json
@@ -1,0 +1,138 @@
+{
+  "main.metadata.description": "Locaplus specializes in canopy and event tent rentals across the Eastern Townships, providing high-quality tents, chairs, tables, and equipment for festivals, weddings, and special events. Serving Bromont, Granby, South Shore, Sherbrooke, and Magog since 1991.",
+
+  "nav.aboutUs": "About Us",
+  "nav.products": "Products",
+  "nav.projects": "Projects",
+  "nav.contactUs": "Contact us",
+  "home.title": "Trusted Tent & Event Rentals in the Eastern Townships",
+  "home.subtitle": "  With 30+ years of experience Locaplus delivers reliable tent event rentals to make your event unforgettable.",
+  "home.requestQuote": "Request a quote",
+
+  "form.firstName": "First name",
+  "form.lastName": "Last name",
+  "form.typeOfEvent": "Type of event",
+  "form.eventDate": "Event date",
+  "form.location": "Location",
+  "form.email": "Email",
+  "form.phoneNumber": "Phone number",
+  "form.message": "Message",
+  "form.submitButton": "Request quote",
+
+  "tent.name.10x10": "10' x 10' Tent",
+  "tent.name.10x20": "10' x 20' Tent",
+  "tent.name.20x20": "20' x 20' Tent",
+  "tent.name.20x30": "20' x 30' Tent",
+  "tent.name.20x40": "20' x 40' Tent",
+  "tent.name.40x40": "40' x 40' Tent",
+  "tent.name.hexagonal": "Hexagonal Tent",
+  "tent.name.40x60": "40' x 60' Tent",
+  "tent.name.40x80": "40' x 80' Tent",
+  "tent.name.40x100": "40' x 100' Tent",
+  "tent.name.40x200": "40' x up to 200' Tent",
+  "tent.name.60x80": "60' x 80' Tent",
+  "tent.name.60x100": "60' x 100' Tent",
+  "tent.name.60x200": "60' x up to 200' Tent",
+  "tents.capacity.range": "Capacity: {min} - {max} people",
+  "tents.capacity.maxOnly": "Capacity up to {max} people",
+
+  "equipment.chairsAndTables": "Chairs and Tables",
+  "equipment.burgundyFoldingChair": "Burgundy folding chair",
+  "equipment.crystalChiavariChair": "Crystal chiavari chair",
+  "equipment.mahoganyChiavariChair": "Mahogany chiavari chair",
+  "equipment.whiteBistroChair": "White bistro chair",
+  "equipment.barStool": "Bar stool",
+  "equipment.cocktailTable30In": "30 in. cocktail table",
+  "equipment.roundTable48In": "48 in. round table",
+  "equipment.roundTable60In": "60 in. round table",
+  "equipment.roundTable72In": "72 in. round table",
+  "equipment.rectangularTable30x96": "30x96 rectangular table",
+  "equipment.rectangularTable30x72": "30x72 rectangular table",
+  "equipment.floorRegular": "Regular floor",
+  "equipment.floorLeveled": "Leveled floor",
+  "equipment.floorDance": "Dance floor",
+
+  "equipment.podium": "Podium",
+  "equipment.redCarpet25Ft": "Red carpet 3 x 25",
+  "equipment.stanchionsAndRopes": "Stanchions and ropes",
+  "equipment.coatRack": "Coat rack",
+  "equipment.equipment": "Equipment",
+  "equipment.cornRoasts": "Corn roasts",
+  "equipment.roastOven": "Roast oven (mechoui)",
+  "equipment.propaneBBQ24x48": "24 x 48 propane BBQ",
+  "equipment.coffeeMaker": "Coffee maker",
+  "equipment.propanePatioHeater": "Propane patio heater",
+  "equipment.portableBar": "Portable bar",
+  "equipment.portableIlluminatedBar": "Portable illuminated bar",
+  "equipment.dishesCutleryAndGlasses": "Dishes, cutlery and glasses",
+  "equipment.iceBasket": "Ice basket",
+  "equipment.propaneFurnace": "Propane furnace",
+  "equipment.propaneFurnaceAccessory": "Propane furnace accessory",
+
+  "about.title": "Transforming Moments into Memories, Rain or Shine",
+  "about.section1.title": "30+ Years of Rental Experience",
+  "about.section1.description": "Locaplus brings decades of expertise to ensure your event runs smoothly, from intimate gatherings to large festivals across Quebec.",
+  "about.section2.title": "Premium Tents & Equipment",
+  "about.section2.description": "Our high-quality tents, tables, chairs, and flooring are designed to elevate any event, rain or shine.",
+  "about.section3.title": "Tailored Event Solutions",
+  "about.section3.description": "We offer customized rentals to match your event's unique needs, creating unforgettable moments every time.",
+  "about.section4.title": "Official distributor and installer for Tempo shelters in the Eastern Townships",
+  "about.section4.description": "Locaplus is the official Tempo dealer in the Eastern Townships, providing sales, distribution, and expert installation.",
+
+  "products.title": "We've got you covered",
+  "products.description": "Premium tents, stylish furniture, and all the essentials to make your event seamless and unforgettable.",
+  "products.tents": "Tents",
+  "products.equipment": "Equipment",
+  "products.flooringChairsAndTables": "Flooring, Chairs and Tables",
+
+  "contact.title": "Have Questions? Ask Us!",
+  "contact.description": "Looking for reliable event rentals? Locaplus offers premium tents and equipment across Bromont, Granby, South Shore, Sherbrooke, and Magog. Reach out or request a quote below!",
+
+  "contactInfo.address": "203 Rue des Alouettes, Saint-Alphonse-de-Granby",
+  "contactInfo.phone": "Phone",
+  "contactInfo.email": "Email",
+  "contactInfo.office": "Office",
+  "contactInfo.followUs": "Follow Us",
+  "contactInfo.450Locations": "Bromont, Granby, South Shore",
+  "contactInfo.819Locations": "Sherbrooke, Magog",
+
+  "requestQuote.title": "Request a quote",
+  "requestQuote.success.title": "Quote Request Submitted",
+  "requestQuote.success.message": "You will receive an automated email confirmation. A team member will follow up with you shortly.",
+  "requestQuote.maintenance": "The quote form will be unavailable until January 5th. Please send all quote requests to contact@locaplus.net. We apologize for the inconvenience.",
+
+  "error.firstName.required": "A first name is required",
+  "error.lastName.required": "A last name is required",
+  "error.eventType.required": "An event type is required",
+  "error.eventDate.required": "An event date is required",
+  "error.dateInput.isValid": "Please enter a valid date",
+  "error.location.required": "A location is required",
+  "error.email.required": "An email is required",
+  "error.email.invalid": "Must be a valid email",
+  "error.phoneNumber.required": "A phone number is required",
+  "error.consentCheckbox.isChecked": "Your consent to the privacy policy is required",
+  "error.sendEmail.generic": "An error occured while sending email",
+  "form.sending": "Sending...",
+
+  "privacyPolicy.checkbox.text": "I have read and agree to the",
+  "privacyPolicy.checkbox.link": "privacy policy",
+
+  "privacyPolicy.title": "Privacy Policy",
+  "privacyPolicy.intro": "At Locaplus, we respect your privacy and are committed to protecting your personal information.",
+
+  "privacyPolicy.infoCollection.title": "Information We Collect",
+  "privacyPolicy.infoCollection.description": "When you fill out the quote request form on our website, we collect the information you provide, such as your name, email address, phone number, and event details.",
+
+  "privacyPolicy.howWeUse.title": "How We Use Your Information",
+  "privacyPolicy.howWeUse.description": "The information you provide is used solely to respond to your request and provide a quote. We do not store your personal information beyond what is necessary to reply to your inquiry.",
+
+  "privacyPolicy.retention.title": "Retention of Information",
+  "privacyPolicy.retention.description": "We retain your personal information only for as long as necessary to fulfill the purpose for which it was collected, or as required by law.",
+
+  "privacyPolicy.protection.title": "Protection of Your Information",
+  "privacyPolicy.protection.description": "We take reasonable steps to ensure the information you provide is securely transmitted.",
+
+  "privacyPolicy.contact.title": "Contact Us",
+  "privacyPolicy.contactInfo": "If you have questions about this Privacy Policy, contact us at",
+  "privacyPolicy.contactEmail": "contact@locaplus.net"
+}
