+++ next.config.ts
@@ -1,5 +1,7 @@
 import type { NextConfig } from 'next';
+import createNextIntlPlugin from 'next-intl/plugin';
 
+const withNextIntl = createNextIntlPlugin();
 const nextConfig: NextConfig = {};
 
-export default nextConfig;
+export default withNextIntl(nextConfig);
