+++ src/i18n/request.ts
@@ -1,0 +1,18 @@
+import { getRequestConfig } from 'next-intl/server';
+
+import { routing } from './routing';
+
+export default getRequestConfig(async ({ requestLocale }) => {
+  // This typically corresponds to the `[locale]` segment
+  let locale = (await requestLocale) as 'en' | 'fr';
+
+  // Ensure that a valid locale is used
+  if (!locale || !routing.locales.includes(locale)) {
+    locale = routing.defaultLocale;
+  }
+
+  return {
+    locale,
+    messages: (await import(`../../messages/${locale}.json`)).default,
+  };
+});
