+++ src/middleware.ts
@@ -1,0 +1,10 @@
+import createMiddleware from 'next-intl/middleware';
+
+import { routing } from './i18n/routing';
+
+export default createMiddleware(routing);
+
+export const config = {
+  // Match only internationalized pathnames
+  matcher: ['/', '/(en|fr)/:path*'],
+};
